+++ Madcamp1st/app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.example.madcamp1st"
+        applicationId "com.example.madcamp2nd"
         minSdkVersion 23
         targetSdkVersion 30
         versionCode 1
@@ -37,4 +37,9 @@ dependencies {
 
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     implementation 'com.karumi:dexter:6.2.2'
+    implementation 'com.facebook.android:facebook-login:[8.1)'
+    implementation 'com.facebook.android:facebook-android-sdk:[4,5)'
+
 }
+//    implementation 'com.facebook.android:facebook-android-sdk:[4,5)'
+//implementation 'com.facebook.android:facebook-android-sdk:[8,9)'
