+++ Madcamp1st/app/src/main/AndroidManifest.xml
@@ -1,11 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.madcamp1st">
+    package="com.example.madcamp2nd">
+
+
+    <queries> <package android:name="com.facebook.katana" /> </queries>
 
     <uses-permission android:name="android.permission.READ_CONTACTS" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.CAMERA"/>
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
+    <uses-permission android:name="android.permission.INTERNET"/>
+
 
     <application
         android:allowBackup="true"
@@ -16,28 +21,27 @@
         android:requestLegacyExternalStorage = "true"
         android:theme="@style/Theme.Madcamp1st">
 
-        <activity android:name=".MainActivity">
+        <activity android:name="com.example.madcamp2nd.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
 
-        <activity android:name=".games.snake.SnakeActivity"
-            android:parentActivityName=".MainActivity"
+        <activity android:name="com.example.madcamp2nd.games.snake.SnakeActivity"
+            android:parentActivityName="com.example.madcamp2nd.MainActivity"
             android:screenOrientation="portrait"/>
 
-        <activity android:name=".games.minesweeper.MinesweeperActivity"
-            android:parentActivityName=".MainActivity"
+        <activity android:name="com.example.madcamp2nd.games.minesweeper.MinesweeperActivity"
+            android:parentActivityName="com.example.madcamp2nd.MainActivity"
             android:screenOrientation="portrait"/>
 
-        <activity android:name=".games.breakout.BreakoutActivity"
-            android:parentActivityName=".MainActivity"
+        <activity android:name="com.example.madcamp2nd.games.breakout.BreakoutActivity"
+            android:parentActivityName="com.example.madcamp2nd.MainActivity"
             android:screenOrientation="landscape"/>
 
         <provider
-            android:authorities="com.example.madcamp1st"
+            android:authorities="com.example.madcamp2nd"
             android:name="androidx.core.content.FileProvider"
             android:exported="false"
             android:requestLegacyExternalStorage = "true"
@@ -48,6 +52,28 @@
                 android:resource="@xml/file_paths" />
 
         </provider>
+
+        <meta-data
+            android:name="com.facebook.sdk.ApplicationId"
+            android:value="@string/facebook_app_id"/>
+            <activity android:name="com.facebook.FacebookActivity"
+                android:configChanges= "keyboard|keyboardHidden|screenLayout|screenSize|orientation"
+                android:label="@string/app_name" />
+            <activity
+                android:name="com.facebook.CustomTabActivity"
+                android:exported="true">
+                <intent-filter>
+                    <action
+                        android:name="android.intent.action.VIEW" />
+                    <category
+                        android:name="android.intent.category.DEFAULT" />
+                    <category
+                        android:name="android.intent.category.BROWSABLE" />
+                    <data
+                        android:scheme="@string/fb_login_protocol_scheme" />
+                </intent-filter>
+            </activity>
+
 
     </application>
