--- Madcamp1st/app/src/main/java/com/example/madcamp1st/contacts/Contact.java
@@ -1,11 +1,0 @@
-package com.example.madcamp1st.contacts;
-
-public class Contact {
-    public String name;
-    public String number;
-
-    public Contact(String name, String number){
-        this.name = name;
-        this.number = number;
-    }
-}
