--- Madcamp1st/app/src/main/java/com/example/madcamp1st/images/Image.java
@@ -1,13 +1,0 @@
-package com.example.madcamp1st.images;
-
-import android.net.Uri;
-
-public class Image {
-    final Uri uri;
-    final String name;
-
-    public Image(Uri uri, String name) {
-        this.uri = uri;
-        this.name = name;
-    }
-}
