+++ Madcamp1st/app/src/main/java/com/example/madcamp2nd/LoginCallback.java
@@ -1,0 +1,57 @@
+package com.example.madcamp2nd;
+
+import android.os.Bundle;
+import android.util.Log;
+
+import com.facebook.AccessToken;
+import com.facebook.FacebookCallback;
+import com.facebook.FacebookException;
+import com.facebook.GraphRequest;
+import com.facebook.GraphResponse;
+import com.facebook.login.LoginResult;
+
+import org.json.JSONObject;
+
+
+public class LoginCallback implements FacebookCallback<LoginResult>
+{
+
+    @Override
+    public void onSuccess(LoginResult loginResult)
+    {
+        Log.e("Callback :: ", "onSuccess");
+        requestMe(loginResult.getAccessToken());
+    }
+
+    @Override
+    public void onCancel()
+    {
+        Log.e("Callback :: ", "onCancel");
+    }
+
+    @Override
+    public void onError(FacebookException error)
+    {
+        Log.e("Callback :: ", "onError : " + error.getMessage());
+    }
+
+    // 사용자 정보 요청
+    public void requestMe(AccessToken token)
+    {
+        GraphRequest graphRequest = GraphRequest.newMeRequest(token,
+                new GraphRequest.GraphJSONObjectCallback()
+                {
+                    @Override
+                    public void onCompleted(JSONObject object, GraphResponse response)
+                    {
+                        Log.e("result",object.toString());
+                    }
+                });
+
+        Bundle parameters = new Bundle();
+        parameters.putString("fields", "id,name,email,gender,birthday");
+        graphRequest.setParameters(parameters);
+        graphRequest.executeAsync();
+    }
+
+}
