+++ Madcamp1st/app/src/main/java/com/example/madcamp2nd/contacts/Contact.java
@@ -1,0 +1,11 @@
+package com.example.madcamp2nd.contacts;
+
+public class Contact {
+    public String name;
+    public String number;
+
+    public Contact(String name, String number){
+        this.name = name;
+        this.number = number;
+    }
+}
