+++ Madcamp1st/app/src/main/java/com/example/madcamp2nd/games/breakout/BreakoutActivity.java
@@ -1,0 +1,73 @@
+package com.example.madcamp2nd.games.breakout;
+
+import android.app.Activity;
+import android.graphics.Point;
+import android.os.Bundle;
+import android.view.Display;
+import android.view.View;
+
+public class BreakoutActivity extends Activity {
+    // gameView will be the view of the game
+    // It will also hold the logic of the game
+    // and respond to screen touches as well
+    private BreakoutEngine breakoutEngine;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        // Get the pixel dimensions of the screen
+        Display display = getDisplay();
+
+        // Initialize the result into a Point object
+        Point size = new Point();
+        display.getRealSize(size);
+
+        // why?
+        size.x -= 112;
+
+        // Initialize gameView and set it as the view
+        breakoutEngine = new BreakoutEngine(this, size);
+        setContentView(breakoutEngine);
+    }
+
+    // This method executes when the player starts the game
+    @Override
+    protected void onResume() {
+        super.onResume();
+        // Tell the gameView resume method to execute
+        breakoutEngine.resume();
+    }
+
+    // This method executes when the player quits the game
+    @Override
+    protected void onPause() {
+        super.onPause();
+        // Tell the gameView pause method to execute
+        breakoutEngine.pause();
+    }
+
+    @Override
+    public void onWindowFocusChanged(boolean hasFocus) {
+        super.onWindowFocusChanged(hasFocus);
+        if (hasFocus) {
+            hideSystemUI();
+        }
+    }
+
+    private void hideSystemUI() {
+        // Enables regular immersive mode.
+        // For "lean back" mode, remove SYSTEM_UI_FLAG_IMMERSIVE.
+        // Or for "sticky immersive," replace it with SYSTEM_UI_FLAG_IMMERSIVE_STICKY
+        getWindow().getDecorView().setSystemUiVisibility(
+                View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
+                        // Set the content to appear under the system bars so that the
+                        // content doesn't resize when the system bars hide and show.
+                        | View.SYSTEM_UI_FLAG_LAYOUT_STABLE
+                        | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
+                        | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
+                        // Hide the nav bar and status bar
+                        | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
+                        | View.SYSTEM_UI_FLAG_FULLSCREEN);
+    }
+}
