+++ Madcamp1st/app/src/main/java/com/example/madcamp2nd/games/minesweeper/views/grid/Cell.java
@@ -1,0 +1,117 @@
+package com.example.madcamp2nd.games.minesweeper.views.grid;
+
+import android.content.Context;
+import android.graphics.Canvas;
+import android.graphics.drawable.Drawable;
+import android.view.View;
+
+import androidx.core.content.ContextCompat;
+
+import com.example.madcamp2nd.R;
+import com.example.madcamp2nd.games.minesweeper.MinesweeperEngine;
+
+public class Cell extends BaseCell implements View.OnClickListener, View.OnLongClickListener {
+    public Cell (Context context, int x, int y){
+        super(context);
+        setPosition(x, y);
+        setOnClickListener(this);
+        setOnLongClickListener(this);
+    }
+
+    @Override
+    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
+        super.onMeasure(widthMeasureSpec, widthMeasureSpec);
+    }
+
+    @Override
+    public void onClick(View v) {
+        MinesweeperEngine.getInstance().click(getXPos(), getYPos());
+    }
+
+    @Override
+    public boolean onLongClick(View v) {
+        if (!isClicked()) MinesweeperEngine.getInstance().flag(getXPos(), getYPos());
+        return true;
+    }
+
+    @Override
+    protected void onDraw(Canvas canvas) {
+        super.onDraw(canvas);
+
+        drawButton(canvas);
+
+        if (isFlagged())
+            drawFlag(canvas);
+        else if (isRevealed() && isBomb() && !isClicked())
+            drawNormalBomb(canvas);
+        else
+            if (isClicked())
+                if(getValue() == -1)
+                    drawBombExploded(canvas);
+                else
+                    drawNumber(canvas);
+            else
+                drawButton(canvas);
+    }
+
+    private void drawBombExploded(Canvas canvas) {
+        Drawable drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_bomb_exploded);
+        drawable.setBounds(0, 0, getWidth(), getHeight());
+        drawable.draw(canvas);
+    }
+
+    private void drawFlag (Canvas canvas) {
+        Drawable drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_flag);
+        drawable.setBounds(0, 0, getWidth(), getHeight());
+        drawable.draw(canvas);
+    }
+
+
+    private void drawButton(Canvas canvas) {
+        Drawable drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_button);
+        drawable.setBounds(0, 0, getWidth(), getHeight());
+        drawable.draw(canvas);
+    }
+
+    private void drawNormalBomb(Canvas canvas) {
+        Drawable drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_bomb_normal);
+        drawable.setBounds(0, 0, getWidth(), getHeight());
+        drawable.draw(canvas);
+    }
+
+    private void drawNumber(Canvas canvas) {
+        Drawable drawable = null;
+        switch (getValue()) {
+            case 0:
+                drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_number_0);
+                break;
+            case 1:
+                drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_number_1);
+                break;
+            case 2:
+                drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_number_2);
+                break;
+            case 3:
+                drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_number_3);
+                break;
+            case 4:
+                drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_number_4);
+                break;
+            case 5:
+                drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_number_5);
+                break;
+            case 6:
+                drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_number_6);
+                break;
+            case 7:
+                drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_number_7);
+                break;
+            case 8:
+                drawable = ContextCompat.getDrawable(getContext(), R.drawable.minesweeper_number_8);
+                break;
+        }
+
+        drawable.setBounds(0, 0, getWidth(), getHeight());
+        drawable.draw(canvas);
+    }
+}
