+++ Madcamp1st/app/src/main/java/com/example/madcamp2nd/games/snake/SnakeActivity.java
@@ -1,0 +1,70 @@
+package com.example.madcamp2nd.games.snake;
+
+import android.app.Activity;
+import android.graphics.Point;
+import android.os.Bundle;
+import android.view.Display;
+import android.view.View;
+
+public class SnakeActivity extends Activity {
+    // Declare an instance of SnakeEngine
+    SnakeEngine snakeEngine;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        // Get the pixel dimensions of the screen
+        Display display = getDisplay();
+
+        // Initialize the result into a Point object
+        Point size = new Point();
+        display.getRealSize(size);
+
+        // why?
+        size.y -= 112;
+
+        // Create a new instance of the SnakeEngine class
+        snakeEngine = new SnakeEngine(this, size);
+        // Make snakeEngine the view of the Activity
+        setContentView(snakeEngine);
+    }
+
+    // Start the thread in snakeEngine
+    @Override
+    protected void onResume() {
+        super.onResume();
+        snakeEngine.resume();
+    }
+
+    // Stop the thread in snakeEngine
+    @Override
+    protected void onPause() {
+        super.onPause();
+        snakeEngine.pause();
+    }
+
+    @Override
+    public void onWindowFocusChanged(boolean hasFocus) {
+        super.onWindowFocusChanged(hasFocus);
+        if (hasFocus) {
+            hideSystemUI();
+        }
+    }
+
+    private void hideSystemUI() {
+        // Enables regular immersive mode.
+        // For "lean back" mode, remove SYSTEM_UI_FLAG_IMMERSIVE.
+        // Or for "sticky immersive," replace it with SYSTEM_UI_FLAG_IMMERSIVE_STICKY
+        getWindow().getDecorView().setSystemUiVisibility(
+                View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
+                        // Set the content to appear under the system bars so that the
+                        // content doesn't resize when the system bars hide and show.
+                        | View.SYSTEM_UI_FLAG_LAYOUT_STABLE
+                        | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
+                        | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
+                        // Hide the nav bar and status bar
+                        | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
+                        | View.SYSTEM_UI_FLAG_FULLSCREEN);
+    }
+}
